feat: new version

+++ post.js
@@ -4,11 +4,19 @@ const fs = require('fs');
 
 try {
   const filename = core.getInput('filename');
-  fs.rmSync(filename, { force: true });
-  // TODO: remove directory if empty
-
-  // eslint-disable-next-line no-console
-  console.log(`${filename} deleted!`);
+  if (filename) {
+    fs.rmSync(filename, { force: true });
+    // eslint-disable-next-line no-console
+    console.log(`${filename} deleted!`);
+    // TODO: remove directory if empty
+  }
+  const filename2 = core.getInput('filename2');
+  if (filename2) {
+    fs.rmSync(filename2, { force: true });
+    // eslint-disable-next-line no-console
+    console.log(`${filename2} deleted!`);
+    // TODO: remove directory if empty
+  }
 } catch (error) {
   core.setFailed(error.message);
 }
